requirements.txt
boto3==1.15.6
scipy==1.4.1
numba==0.51.2
astroid==2.4.1
chardet==3.0.4
cycler==0.10.0
idna==2.9
isort==4.3.21
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
gspread==3.6.0
gspread-formatting==0.2.5
GitPython==3.1.0
numpy==1.18.2
matplotlib==3.2.1
mccabe==0.6.1
numpy==1.18.2
pandas==1.0.3
pyshp==2.1.0
pymc3==3.8
requests==2.23.0
jupyter==1.0.0
jupyterlab==2.1.4
simplejson==3.17.0
ujson==2.0.3
seaborn==0.10.0
click==7.1.1
sentry-sdk==0.14.3
pandarallel==1.4.8
scikit-learn==0.22.2.post1
pip
openapi-schema-pydantic==1.1.0
iminuit==1.3.10
fastparquet==0.3.3
us==1.0.0
adtk==0.6.0
-e .
pydantic==1.7.3
dill
dictdiffer==0.8.1
structlog==20.1.0
structlog-sentry==1.2.2
more_itertools==8.3.0
typing-extensions==3.7.4.3
xlrd==1.2.0  # For panda's read_excel method
compress_pickle==2.0.1

# Dependencies recommended by pandas for better performance
numexpr==2.7.1

backports.cached-property==1.0.0.post2

python-snappy==0.6.0


# Mostly for dash_app.py
Brotli==1.0.7
certifi==2020.6.20
dash==1.14.0
dash-core-components==1.10.2
dash-html-components==1.0.3
dash-renderer==1.6.0
dash-table==4.9.0
dash-pivottable==0.0.2
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
plotly==4.9.0
PyYAML==5.3.1
retrying==1.3.3
six==1.15.0
Werkzeug==1.0.1
